source/shared/infrastructure/requirements.txt
aws-cdk.core>=1.85.0,<2.0
aws-cdk.custom-resources>=1.85.0,<2.0
aws-cdk.aws-dynamodb>=1.85.0,<2.0
aws-cdk.aws-events>=1.85.0,<2.0
aws-cdk.aws-events-targets>=1.85.0,<2.0
aws-cdk.aws-iam>=1.85.0,<2.0
aws-cdk.aws-lambda>=1.85.0,<2.0
aws-cdk.aws-s3>=1.85.0,<2.0
aws-cdk.aws-s3-deployment>=1.85.0,<2.0
aws-cdk.aws-sqs>=1.85.0,<2.0
aws-cdk.aws-ssm>=1.85.0,<2.0
aws-cdk.aws-cloudfront>=1.85.0,<2.0
aws-cdk.aws-cloudfront-origins>=1.85.0,<2.0
